pyproject.toml
[tool.poetry]
name = "pythonspain"
version = "1.0.0"
description = "A backend for Python Spain"
authors = ["Python Spain <contacto@es.python.org>"]

[tool.poetry.dependencies]
python = "^3.7"
bleach = "^3.1.0"
redis = "^3.4.1"
django = "^3.0.3"
celery = "^4.4.0"
django-extensions = "^2.2.8"
django-environ = "^0.4.5"
django-anymail = "^7.0.0"
Pillow = "^7.0.0"
django-snitch = "^1.5.0"
django-filter = "^2.2.0"
django-redis = "^4.11.0"
django-model-utils = "^4.0.0"
django-simple-options = "^2.1.1"
djangorestframework = "^3.11.0"
coreapi = "^2.3.3"
argon2-cffi = "^19.2.0"
psycopg2-binary = "^2.8.4"
Sphinx = "^2.4.1"
sphinx-rtd-theme = "^0.4.3"
sphinxcontrib-httpdomain = "^1.7.0"
raven = "^6.10.0"
gunicorn = "^20.0.4"
tornado = "<6.0.0,>=4.2.0"
xlrd = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "<5.4.0"
django-debug-toolbar = "^2.2"
factory_boy = "^2.12.0"
django-test-plus = "^1.4.0"
mypy = "^0.761"
pytest-sugar = "^0.9.2"
flake8 = "^3.7.9"
coverage = "^5.0.3"
black = "^19.10b0"
pytest-django = "^3.8.0"
Werkzeug = "^1.0.0"
django_coverage_plugin = "^1.8.0"
pylint-django = "^2.0.13"
pylint-celery = "^0.3"
pre-commit = "^2.0.1"
ipython = "^7.13.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
